Fulfill the suggestions of Ulf
